refactor: move logic in main and main.test to Book and Book.test
