--- src/main.test.ts
@@ -1,7 +1,0 @@
-import Book from "./main";
-import { expect, test } from "vitest";
-
-test("default constructed values", () => {
-    const agot = new Book("A Game of Thrones", "George R. R. Martin", 694, true);
-    expect(agot.info()).toBe("A Game of Thrones by George R. R. Martin, 694 pages, read");
-});
